Initial refactor of syncer and log_sync.
